+++ src/js/utils/helpers/CircleHelper.d.ts
@@ -8,4 +8,18 @@
  * @modified 2020-10-18 Ported to Typescript from vanilla JS.
  * @version  1.0.0
  **/
-export {};
+import { Circle } from "../../Circle";
+import { PlotBoilerplate } from "../../PlotBoilerplate";
+import { Vertex } from "../../Vertex";
+export declare class CircleHelper {
+    /**
+     * The constructor.
+     *
+     * @constructor
+     * @name VoronoiCell
+     * @param {Circle} circle - The circle to handle.
+     * @param {Vertex} radiusPoint - A point to define the radius (distance from center).
+     * @param {PlotBoilerplate} pb - The PlotBoilerplate which contains the circle and point.
+     **/
+    constructor(circle: Circle, radiusPoint: Vertex, pb: PlotBoilerplate);
+}